requirements.txt
apispec==0.20.0
click==6.7
Flask==0.12
Flask-Cors==3.0.2
gevent==1.2.1
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
marshmallow==2.13.4
marshmallow-jsonschema==0.3.0
PyJWT==1.4.2
PyMySQL==0.7.10
PyYAML==3.12
requests==2.13.0
six==1.10.0
SQLAlchemy==1.1.6
Werkzeug==0.12
xpinyin==0.5.5
Flask-Cache==0.13.1
itsdangerous==0.24
validators==0.12.0
elasticsearch==5.4.0
traceback2==1.4.0
Flask-RESTful==0.3.6
boto
apscheduler
python-dateutil
------------------------
dockerpty
pytest
requests
